requirements.txt
scipy==1.5.4
numpy==1.18.5
matplotlib==3.3.2
opencv_python==3.4.8.29
tensorflow_gpu==2.3.1
Pillow==8.0.1
PyYAML==5.3.1
tensorflow==2.3.1
